fix: start dock with current config
